test: move tests to source folder
specifying test/bla.test.ts is kinda pointless.
Move them to the sources and name them source/bla.test.ts is
not duplicating the test in the file path and is autodetected.
