+++ .github/workflows/build.yml
@@ -25,7 +25,7 @@ jobs:
         run: deno task check
 
       - name: Test
-        run: deno task test
+        run: deno test
 
   backport:
     runs-on: ubuntu-latest
