+++ deno.jsonc
@@ -2,7 +2,6 @@
   "lock": false,
   "tasks": {
     "check": "deno cache --check=all source/mod.ts",
-    "test": "deno test test/",
     "backport": "deno run --no-prompt --allow-read=. --allow-write=. https://deno.land/x/deno2node@v1.8.1/src/cli.ts tsconfig.json"
   },
   "fmt": {
